routes.py addition, trimming of main.py, added .gitignore, adjusted readme.py
