+++ .gitignore
@@ -1,0 +1,13 @@
+# Python
+__pycache__/
+*.py[cod]
+*.egg-info/
+.env
+.venv/
+
+# Logs
+*.log
+
+# Databases
+*.sqlite3
+*.db
