+++ README.md
@@ -1,79 +1,79 @@
+# Student Management System
+
 Core Modules to Include:
-
-Authentication & Authorization: Admin, Faculty, Students
-
-Student Management: Enrollments, attendance, transcripts
-
-Course Management: Add/edit/delete courses, assign faculty
-
-Faculty Management: Assign courses, view student lists
-
-Timetable/Schedule Management
-
-Grades & Results
-
-Fees & Payments
-
-Notifications & Messaging
-
-Reports & Analytics (optional)
+* Authentication & Authorization: Admin, Faculty, Students
+* Student Management: Enrollments, attendance, transcripts
+* Course Management: Add/edit/delete courses, assign faculty
+* Faculty Management: Assign courses, view student lists
+* Timetable/Schedule Management
+* Grades & Results
+* Fees & Payments
+* Notifications & Messaging
+* Reports & Analytics (optional)
 
 2. Plan the Data Model (Database Design)
 
 Use an ER diagram to define entities and relationships. Example tables:
+* Users: id, name, email, password, role
+* Students: student_id, user_id, dob, department_id
+* Faculty: faculty_id, user_id, department_id
+* Departments: id, name
+* Courses: id, name, credits, department_id
+* Enrollments: id, student_id, course_id, semester
+* Grades: id, enrollment_id, grade
+* Schedules: id, course_id, faculty_id, day, time
+* Payments: id, student_id, amount, status, date
+* Use SQLite for development (or PostgreSQL/MySQL for deployment).
 
-Users: id, name, email, password, role
-
-Students: student_id, user_id, dob, department_id
-
-Faculty: faculty_id, user_id, department_id
-
-Departments: id, name
-
-Courses: id, name, credits, department_id
-Enrollments: id, student_id, course_id, semester
-Grades: id, enrollment_id, grade
-Schedules: id, course_id, faculty_id, day, time
-Payments: id, student_id, amount, status, date
-Use SQLite for development (or PostgreSQL/MySQL for deployment).
 4. Setup the Project Structure
- 
- 
- 
- 
- 
- 
- 
- 
+
 🛠️ 4. Tech Stack Setup
-Use Flask for the backend.
-Flask-SQLAlchemy for ORM
-Flask-Migrate for database migrations
-Jinja2 for templating
-Optional: Flask-Login for user authentication
+
+* Use Flask for the backend.
+* Flask-SQLAlchemy for ORM
+* Flask-Migrate for database migrations
+* Jinja2 for templating
+* Optional: Flask-Login for user authentication
+
 🗂️ 5. Implement in Phases
+
 📌 Phase 1: Setup
-Initialize Flask app and configure SQLAlchemy
-Create user login & role-based access
+
+* Initialize Flask app and configure SQLAlchemy
+* Create user login & role-based access
+
 📌 Phase 2: Core Features
-Student CRUD
-Course CRUD
-Faculty CRUD
-Enrollment system
-Grade entry/view
-📌 Phase 3: Scheduling & Timetables
-Course schedule creation
-Faculty & student views
-📌 Phase 4: Fees & Payments
-Payment tracking
-Payment status views
-📌 Phase 5: Reports
-Student transcripts
-Course-wise grade reports
+
+* Student CRUD
+* Course CRUD
+* Faculty CRUD
+* Enrollment system
+* Grade entry/view
+
+📌 Phase 3:
+* Scheduling & Timetables
+* Course schedule creation
+* Faculty & student views
+
+📌 Phase 4:
+
+* Fees & Payments
+* Payment tracking
+* Payment status views
+
+📌 Phase 5:
+
+* Reports
+* Student transcripts
+* Course-wise grade reports
+
 🧪 6. Testing
-Unit testing using pytest
-Manual test cases for all forms, role-based access, and data flows
+
+* Unit testing using pytest
+* Manual test cases for all forms, role-based access, and data flows
+
 🚀 7. Deployment (Optional)
-Use Heroku or PythonAnywhere for deployment
-Include .env for secrets and config
-Use gunicorn + nginx if deploying on a VPS
+
+* Use Heroku or PythonAnywhere for deployment
+* Include .env for secrets and config
+* Use gunicorn + nginx if deploying on a VPS
